Refactore PropsEditor component
Issue: #7
Card: S01E06
